requirements.txt
Flask>=3.0.0
pandas>=2.2.0
msal>=1.25.0
requests>=2.31.0
pystray>=0.19.5
Pillow>=10.1.0
cryptography>=41.0.7
pyinstaller>=6.3.0
python-docx>=1.1.0
openpyxl>=3.1.2
